DBクライアントの改善

+++ db/client.py
@@ -1,7 +1,9 @@
+from enum import Enum
 from time import sleep
 from typing import Any
 
 import meilisearch
+from meilisearch.models.task import TaskInfo
 
 from db.settings import Settings
 
@@ -21,29 +23,60 @@ class DBClient:
         self._client = meilisearch.Client(url=self._uri)
 
     def create_index(self, index_name: str) -> None:
+        if self._exists_index(index_name=index_name):
+            raise IndexAlreadyExistsError
         self._client.create_index(uid=index_name)
-        # TODO: すでにインデックスが存在する場合の警告
 
     def delete_index(self, index_name: str) -> None:
+        if not self._exists_index(index_name=index_name):
+            raise IndexNotExistsError
         self._client.delete_index(uid=index_name)
-        # TODO: インデックスが存在しない場合の警告
 
-    def add_document(self, index_name: str, document: Document) -> None:
+    def _exists_index(self, index_name: str) -> bool:
+        indexes = self._client.get_indexes()["results"]
+        if index_name in [index.uid for index in indexes]:
+            return True
+        return False
+
+    def add_document(self, index_name: str, key: str, document: Document) -> None:
         index = self._client.index(uid=index_name)
-        # TODO: URLのユニークチェック
-        task_id = index.add_documents(documents=[document]).task_uid
 
+        documents = self.search_documents(index_name=index_name, attributes=[key], keyword=f'"{document[key]}"')
+        if documents:
+            raise URLAlreadyExistsError
+
+        task = index.add_documents(documents=[document])
+        self._check_task_status(index_name=index_name, task=task)
+
+    def _check_task_status(self, index_name: str, task: TaskInfo) -> None:
         task_status = None
-        while task_status != "succeeded":
+        while task_status != TaskStatus.Succeeded:
             sleep(1)
-            task_status = index.get_task(uid=task_id).status
+            task_status = self._client.index(uid=index_name).get_task(uid=task.task_uid).status
 
-            if task_status == "failed":
+            if task_status == TaskStatus.Failed:
                 raise InvalidDocumentError
 
-    def search_documents(self, index_name: str, keyword: str) -> Documents:
-        documents = self._client.index(uid=index_name).search(keyword, {"attributesToHighlight": ["title", "url"]})
+    def search_documents(self, index_name: str, attributes: list[str], keyword: str) -> Documents:
+        documents = self._client.index(uid=index_name).search(keyword, {"attributesToHighlight": attributes})
         return documents["hits"]
+
+
+class TaskStatus(str, Enum):
+    Succeeded = "succeeded"
+    Failed = "failed"
+
+
+class IndexAlreadyExistsError(Exception):
+    def __init__(self) -> None:
+        super().__init__()
+        self.message = "Index already exists"
+
+
+class IndexNotExistsError(Exception):
+    def __init__(self) -> None:
+        super().__init__()
+        self.message = "Index not exists"
 
 
 class URLAlreadyExistsError(Exception):
